refactor(compiler-core): centralize compiler options
