+++ packages/compiler-core/src/index.ts
@@ -1,6 +1,7 @@
-import { parse, ParserOptions } from './parse'
-import { transform, TransformOptions } from './transform'
-import { generate, CodegenOptions, CodegenResult } from './codegen'
+import { CompilerOptions } from './options'
+import { parse } from './parse'
+import { transform } from './transform'
+import { generate, CodegenResult } from './codegen'
 import { RootNode } from './ast'
 import { isString } from '@vue/shared'
 import { transformIf } from './transforms/vIf'
@@ -15,8 +16,6 @@ import { trackSlotScopes, trackVForSlotScopes } from './transforms/vSlot'
 import { transformText } from './transforms/transformText'
 import { transformOnce } from './transforms/vOnce'
 import { transformModel } from './transforms/vModel'
-
-export type CompilerOptions = ParserOptions & TransformOptions & CodegenOptions
 
 // we name it `baseCompile` so that higher order compilers like @vue/compiler-dom
 // can export `compile` while re-exporting everything else.
@@ -75,22 +74,22 @@ export function baseCompile(
 }
 
 // Also expose lower level APIs & types
-export { parse, ParserOptions, TextModes } from './parse'
+export {
+  CompilerOptions,
+  ParserOptions,
+  TransformOptions,
+  CodegenOptions
+} from './options'
+export { parse, TextModes } from './parse'
 export {
   transform,
   createStructuralDirectiveTransform,
-  TransformOptions,
   TransformContext,
   NodeTransform,
   StructuralDirectiveTransform,
   DirectiveTransform
 } from './transform'
-export {
-  generate,
-  CodegenOptions,
-  CodegenContext,
-  CodegenResult
-} from './codegen'
+export { generate, CodegenContext, CodegenResult } from './codegen'
 export {
   ErrorCodes,
   CoreCompilerError,
